Separate action generator from test suite
