--- monitoring/uss_qualifier/suites/interuss/generate_test_data.yaml
@@ -1,19 +1,0 @@
-name: Generate RID test data
-resources:
-  adjacent_circular_flights_data: resources.netrid.FlightDataResource
-  adjacent_circular_storage_config: resources.netrid.FlightDataStorageResource
-  kml_flights_data: resources.netrid.FlightDataResource
-  kml_storage_config: resources.netrid.FlightDataStorageResource
-actions:
-- test_scenario:
-    scenario_type: scenarios.astm.netrid.StoreFlightData
-    resources:
-      flights_data: adjacent_circular_flights_data
-      storage_configuration: adjacent_circular_storage_config
-  on_failure: Continue
-- test_scenario:
-    scenario_type: scenarios.astm.netrid.StoreFlightData
-    resources:
-      flights_data: kml_flights_data
-      storage_configuration: kml_storage_config
-  on_failure: Continue
