--- monitoring/uss_qualifier/suites/interuss/generate_test_data_twice.yaml
@@ -1,26 +1,0 @@
-name: Generate RID test data twice
-resources:
-  adjacent_circular_flights_data: resources.netrid.FlightDataResource
-  adjacent_circular_storage_config: resources.netrid.FlightDataStorageResource
-  kml_flights_data: resources.netrid.FlightDataResource
-  kml_storage_config: resources.netrid.FlightDataStorageResource
-actions:
-- action_generator:
-    generator_type: action_generators.Repeat
-    specification:
-      action_to_repeat:
-        test_suite:
-          suite_type: suites.interuss.generate_test_data
-          resources:
-            adjacent_circular_flights_data: adjacent_circular_flights_data
-            adjacent_circular_storage_config: adjacent_circular_storage_config
-            kml_flights_data: kml_flights_data
-            kml_storage_config: kml_storage_config
-        on_failure: Abort
-      times_to_repeat: 2
-    resources:
-      adjacent_circular_flights_data: adjacent_circular_flights_data
-      adjacent_circular_storage_config: adjacent_circular_storage_config
-      kml_flights_data: kml_flights_data
-      kml_storage_config: kml_storage_config
-  on_failure: Continue
