+++ schemas/monitoring/uss_qualifier/action_generators/definitions/ActionGeneratorDefinition.json
@@ -1,0 +1,37 @@
+{
+  "$id": "https://github.com/interuss/monitoring/blob/main/schemas/monitoring/uss_qualifier/action_generators/definitions/ActionGeneratorDefinition.json",
+  "$schema": "https://json-schema.org/draft/2020-12/schema",
+  "description": "monitoring.uss_qualifier.action_generators.definitions.ActionGeneratorDefinition, as defined in monitoring/uss_qualifier/action_generators/definitions.py",
+  "properties": {
+    "$ref": {
+      "description": "Path to content that replaces the $ref",
+      "type": "string"
+    },
+    "generator_type": {
+      "description": "Type of action generator",
+      "type": "string"
+    },
+    "resources": {
+      "additionalProperties": {
+        "type": "string"
+      },
+      "description": "Mapping of the ID a resource will be known by in the child action -> the ID a resource is known by in the parent test suite.\n\nThe child action resource ID <key> is supplied by the parent test suite resource ID <value>.\n\nResources not included in this field will not be available to the child action.\n\nIf the parent resource ID is suffixed with ? then the resource will not be required (and will not be populated for the child action when not present in the parent)",
+      "properties": {
+        "$ref": {
+          "description": "Path to content that replaces the $ref",
+          "type": "string"
+        }
+      },
+      "type": "object"
+    },
+    "specification": {
+      "description": "Specification of action generator; format is the ActionGeneratorSpecificationType that corresponds to the `generator_type`",
+      "type": "object"
+    }
+  },
+  "required": [
+    "generator_type",
+    "resources"
+  ],
+  "type": "object"
+}
