--- schemas/monitoring/uss_qualifier/suites/definitions/ActionGeneratorDefinition.json
@@ -1,37 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "type": "object",
-  "properties": {
-    "$ref": {
-      "type": "string",
-      "description": "Path to content that replaces the $ref"
-    },
-    "generator_type": {
-      "type": "string",
-      "description": "Type of action generator"
-    },
-    "resources": {
-      "type": "object",
-      "properties": {
-        "$ref": {
-          "type": "string",
-          "description": "Path to content that replaces the $ref"
-        }
-      },
-      "additionalProperties": {
-        "type": "string"
-      },
-      "description": "Mapping of the ID a resource will be known by in the child action -> the ID a resource is known by in the parent test suite.\n\nThe child action resource ID <key> is supplied by the parent test suite resource ID <value>.\n\nResources not included in this field will not be available to the child action.\n\nIf the parent resource ID is suffixed with ? then the resource will not be required (and will not be populated for the child action when not present in the parent)"
-    },
-    "specification": {
-      "type": "object",
-      "description": "Specification of action generator; format is the ActionGeneratorSpecificationType that corresponds to the `generator_type`"
-    }
-  },
-  "$id": "https://github.com/interuss/monitoring/blob/main/schemas/monitoring/uss_qualifier/suites/definitions/ActionGeneratorDefinition.json",
-  "description": "monitoring.uss_qualifier.suites.definitions.ActionGeneratorDefinition, as defined in monitoring/uss_qualifier/suites/definitions.py",
-  "required": [
-    "generator_type",
-    "resources"
-  ]
-}
